requirements.txt
azureml_automl_core==1.50.0
azureml_core==1.50.0
azureml_dataprep_native==38.0.0
azureml_pipeline_core==1.50.0.post1
azureml_pipeline_steps==1.50.0
azureml_telemetry==1.50.0
azureml_train_automl_client==1.50.0
azureml_train_core==1.50.0
azureml_train_restclients_hyperdrive==1.50.0
nltk==3.8.1
pandas==1.5.3
scikit_learn==1.1.3
azureml-defaults
